+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 ## 1.7 branch
 # 1.7.0
 * `ThumberShell` has been replaced with console commands. Every method of the
-    previous class is now a `Command` class.
+    previous class is now a `Command` class;
+* `ConsoleIntegrationTestCase` and `IntegrationTestCase` have been replaced by
+    `ConsoleIntegrationTestTrait` and `IntegrationTestTrait`;
+* `TestCaseTrait` has been removed and its methods moved to `TestCase`.
 
 ## 1.6 branch
 # 1.6.1